+++ src/Basket/BasketService.php
@@ -423,7 +423,8 @@ final class BasketService extends AbstractService
     }
 
     /**
-     * Sets a pickup point as the basket's shipping destination.
+     * Sets a pickup point as the basket's shipping destination for a
+     * Chrono Relais shipping type.
      *
      * @param SetPickupPointCommand $command
      * @throws SomeParametersAreInvalid
@@ -431,7 +432,32 @@ final class BasketService extends AbstractService
     public function setPickupPoint(SetPickupPointCommand $command): void
     {
         $command->validate();
+
         $this->client->post('basket/'.$command->getBasketId().'/chronorelais-pickup-point', [
+            RequestOptions::JSON => [
+                'pickupPointId' => $command->getPickupPointId(),
+                'title' => $command->getTitle()->getValue(),
+                'firstName' => $command->getFirstName(),
+                'lastName' => $command->getLastName(),
+            ],
+        ]);
+    }
+
+    /**
+     * Sets a pickup point as the basket's shipping destination for a
+     * Mondial Relay shipping type.
+     *
+     * @param SetPickupPointCommand $command
+     *
+     * @return array The full address
+     *
+     * @throws SomeParametersAreInvalid
+     */
+    public function setMondialRelayPickupPoint(SetPickupPointCommand $command): array
+    {
+        $command->validate();
+
+        return $this->client->post(sprintf('basket/%s/mondialrelay-pickup-point', $command->getBasketId()), [
             RequestOptions::JSON => [
                 'pickupPointId' => $command->getPickupPointId(),
                 'title' => $command->getTitle()->getValue(),