+++ src/Organisation/OrganisationBasket.php
@@ -30,7 +30,17 @@ class OrganisationBasket
     /**
      * @var boolean
      */
+    private $checkout;
+
+    /**
+     * @var boolean
+     */
     private $accepted;
+
+    /**
+     * @var boolean
+     */
+    private $hidden;
 
     /**
      * @var \DateTime
@@ -43,7 +53,9 @@ class OrganisationBasket
         $this->setUserId($data['userId']);
         $this->setName($data['name']);
         $this->setLocked($data['locked']);
+        $this->setCheckout($data['checkout']);
         $this->setAccepted($data['accepted']);
+        $this->setHidden($data['hidden']);
         $this->setCreatedAt(\DateTime::createFromFormat(\DateTime::RFC3339, $data['createdAt']));
     }
 
@@ -109,6 +121,38 @@ class OrganisationBasket
     public function setAccepted(bool $accepted): void
     {
         $this->accepted = $accepted;
+    }
+
+    /**
+     * @return bool
+     */
+    public function isCheckout(): bool
+    {
+        return $this->checkout;
+    }
+
+    /**
+     * @param bool $checkout
+     */
+    public function setCheckout(bool $checkout): void
+    {
+        $this->checkout = $checkout;
+    }
+
+    /**
+     * @return bool
+     */
+    public function isHidden(): bool
+    {
+        return $this->hidden;
+    }
+
+    /**
+     * @param bool $hidden
+     */
+    public function setHidden(bool $hidden): void
+    {
+        $this->hidden = $hidden;
     }
 
     /**