+++ tests/User/UserServiceTest.php
@@ -20,6 +20,7 @@ use Wizaplace\SDK\User\UserAddress;
 use Wizaplace\SDK\User\UserAlreadyExists;
 use Wizaplace\SDK\User\UserService;
 use Wizaplace\SDK\User\UserTitle;
+use Wizaplace\SDK\User\UserType;
 
 /**
  * @see UserService
@@ -52,6 +53,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame(null, $user->getBirthday());
         $this->assertNull($user->getCompanyId());
         $this->assertFalse($user->isVendor());
+        $this->assertSame(UserType::CLIENT()->getValue(), $user->getType()->getValue());
 
         // shipping address
         $this->assertNull($user->getShippingAddress()->getTitle());
@@ -130,6 +132,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame(null, $user->getBirthday());
         $this->assertNull($user->getCompanyId());
         $this->assertFalse($user->isVendor());
+        $this->assertSame(UserType::CLIENT()->getValue(), $user->getType()->getValue());
 
         // shipping address
         $this->assertSame($userShipping->getTitle()->getValue(), $user->getShippingAddress()->getTitle()->getValue());
@@ -251,6 +254,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame(null, $user->getBirthday());
         $this->assertNull($user->getCompanyId());
         $this->assertFalse($user->isVendor());
+        $this->assertSame(UserType::CLIENT()->getValue(), $user->getType()->getValue());
 
         //shipping adress
         $this->assertSame($userShipping->getTitle()->getValue(), $user->getShippingAddress()->getTitle()->getValue());
@@ -323,6 +327,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame('1998-07-12', $user->getBirthday()->format('Y-m-d'));
         $this->assertNull($user->getCompanyId());
         $this->assertFalse($user->isVendor());
+        $this->assertSame(UserType::CLIENT()->getValue(), $user->getType()->getValue());
 
         // shipping address
         $this->assertTrue($addressCommand->getTitle()->equals($user->getShippingAddress()->getTitle()));
@@ -376,7 +381,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame(null, $user->getBirthday());
         $this->assertNull($user->getCompanyId());
         $this->assertFalse($user->isVendor());
-
+        $this->assertSame(UserType::CLIENT()->getValue(), $user->getType()->getValue());
 
         $userService->updateUser(
             (new UpdateUserCommand())
@@ -398,6 +403,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame('1963-02-17', $user->getBirthday()->format('Y-m-d'));
         $this->assertNull($user->getCompanyId());
         $this->assertFalse($user->isVendor());
+        $this->assertSame(UserType::CLIENT()->getValue(), $user->getType()->getValue());
     }
 
     public function testUpdateUserAddresses()
@@ -650,6 +656,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame('Jean', $user->getFirstname());
         $this->assertSame('Paul', $user->getLastname());
         $this->assertSame(null, $user->getBirthday());
+        $this->assertSame(UserType::CLIENT()->getValue(), $user->getType()->getValue());
 
         $userService->updateUser(
             (new UpdateUserCommand())
@@ -667,6 +674,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame('Jacques', $user->getFirstname());
         $this->assertSame('Jules', $user->getLastname());
         $this->assertNull($user->getBirthday());
+        $this->assertSame(UserType::CLIENT()->getValue(), $user->getType()->getValue());
     }
 
     public function testRecoverPassword()