+++ CHANGELOG.md
@@ -1,12 +1,33 @@
 # Changelog
 
-## Next release 
+## Next Release
 
-- Added `\Wizaplace\SDK\Order\OrderItem::getProductImageId`
+### New features
+
+- Added `Wizaplace\SDK\Basket\BasketService::setMondialRelayPickupPoint`
+- Added `Wizaplace\SDK\Shipping\MondialRelayService` and its associated classes
+- Added `Wizaplace\SDK\Vendor\Order\OrderService::generateMondialRelayLabel`
+- Added `\Wizaplace\SDK\Order\OrganisationOrderService::getOrder`
+- Added attribute `hidden` to `\Wizaplace\SDK\Organisation\OrganisationBasket`
 - Updated `\Wizaplace\SDK\Catalog\Product` with `\Wizaplace\SDK\Division\Division` 
 - Added `\Wizaplace\SDK\Division\DivisionService::get`
-- Updated `\Wizaplace\SDK\Organisation\OrganisationService::getOrganisationOrders`
-- Updated `tests/Organisation/OrganisationServiceTest`
+
+## 1.41.0
+
+Compatible with Wizaplace 1.27.0
+
+### New requirements
+
+- "psr/log": "^1.0"
+
+### New features
+
+- Added optional parameter `$requestLogger` to the constructor of `\Wizaplace\SDK\ApiClient`
+- Added `\Wizaplace\SDK\Order\OrderItem::getProductImageId`
+- Added `\Wizaplace\SDK\User\User::getType`
+- Added enum `\Wizaplace\SDK\User\UserType`
+- Added pagination data to the array returned by `\Wizaplace\SDK\Organisation\OrganisationService::getOrganisationOrders`
+- Added `\Wizaplace\SDK\Organisation\OrganisationBasket::isCheckout`
 
 ## 1.40.0
 
@@ -52,7 +73,7 @@
 ### New features
 
 - Added optional parameters `$billing` and `$shipping` to `\Wizaplace\SDK\User\UserService::register`
-- Added `\Wizaplace\SDK\Order\Order::getCompanyName` 
+- Added `\Wizaplace\SDK\Order\Order::getCompanyName`
 
 ### Corrections